change update of widget daily on midnight, split Extensions of App into multiple files to reuse parts in widget, split MensaService and CafeteriaLocation to unique files
