+++ TUM Campus App/Extensions/DateFormatter+Extensions.swift
@@ -1,0 +1,33 @@
+//
+//  DateFormatter+Extensions.swift
+//  TUM Campus App
+//
+//  Created by Nikolai Madlener on 12.12.21.
+//  Copyright © 2021 TUM. All rights reserved.
+//
+
+import Foundation
+
+extension DateFormatter {
+    /// yyyy-MM-dd HH:mm:ss
+    static let yyyyMMddhhmmss: DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.timeZone = TimeZone(identifier: "Europe/Berlin")!
+        formatter.dateFormat = "yyyy-MM-dd HH:mm:ss"
+        return formatter
+    }()
+    /// "yyyy-MM-dd"
+    static let yyyyMMdd: DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.timeZone = TimeZone(identifier: "Europe/Berlin")!
+        formatter.dateFormat = "yyyy-MM-dd"
+        return formatter
+    }()
+    /// "EEEE, dd. MMM"
+    static let EEEEddMMM: DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.timeZone = TimeZone(identifier: "Europe/Berlin")!
+        formatter.dateFormat = "EEEE, dd. MMM"
+        return formatter
+    }()
+}
