+++ TUM Campus App/Extensions/Extensions.swift
@@ -93,23 +93,6 @@ extension CodingUserInfoKey {
     static let context = CodingUserInfoKey(rawValue: "context")!
 }
 
-extension DateFormatter {
-    /// yyyy-MM-dd HH:mm:ss
-    static let yyyyMMddhhmmss: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.timeZone = TimeZone(identifier: "Europe/Berlin")!
-        formatter.dateFormat = "yyyy-MM-dd HH:mm:ss"
-        return formatter
-    }()
-    /// "yyyy-MM-dd"
-    static let yyyyMMdd: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.timeZone = TimeZone(identifier: "Europe/Berlin")!
-        formatter.dateFormat = "yyyy-MM-dd"
-        return formatter
-    }()
-}
-
 extension Session {
     static let defaultSession: Session = {
         let adapterAndRetrier = Interceptor(adapter: AuthenticationHandler(), retrier: AuthenticationHandler())
@@ -184,90 +167,12 @@ extension CLLocationCoordinate2D  {
     var location: CLLocation { CLLocation(latitude: latitude, longitude: longitude) }
 }
 
-
-extension Date {
-    var calendar: Calendar { Calendar(identifier: Calendar.current.identifier) }
-    var weekOfMonth: Int { calendar.component(.weekOfMonth, from: self) }
-    
-    var isoCalendar: Calendar { Calendar(identifier: .iso8601) }
-    var weekOfYear: Int { isoCalendar.component(.weekOfYear, from: self) }
-
-    var year: Int {
-        get {
-            return calendar.component(.year, from: self)
-        }
-        set {
-            guard newValue > 0 else { return }
-            let currentYear = calendar.component(.year, from: self)
-            let yearsToAdd = newValue - currentYear
-            if let date = calendar.date(byAdding: .year, value: yearsToAdd, to: self) {
-                self = date
-            }
-        }
-    }
-
-    var month: Int {
-        get {
-            return calendar.component(.month, from: self)
-        }
-        set {
-            let allowedRange = calendar.range(of: .month, in: .year, for: self)!
-            guard allowedRange.contains(newValue) else { return }
-
-            let currentMonth = calendar.component(.month, from: self)
-            let monthsToAdd = newValue - currentMonth
-            if let date = calendar.date(byAdding: .month, value: monthsToAdd, to: self) {
-                self = date
-            }
-        }
-    }
-
-    var day: Int {
-        get {
-            return calendar.component(.day, from: self)
-        }
-        set {
-            let allowedRange = calendar.range(of: .day, in: .month, for: self)!
-            guard allowedRange.contains(newValue) else { return }
-
-            let currentDay = calendar.component(.day, from: self)
-            let daysToAdd = newValue - currentDay
-            if let date = calendar.date(byAdding: .day, value: daysToAdd, to: self) {
-                self = date
-            }
-        }
-    }
-
-    func isLaterThanOrEqual(to date: Date) -> Bool {
-        return self.compare(date) == .orderedDescending || self.compare(date) == .orderedSame
-    }
-
-    /// Determine if date is within the current day
-    var isToday: Bool {
-        let calendar = Calendar.autoupdatingCurrent
-        return calendar.isDateInToday(self)
-    }
-
-    /// Determine if date is within the day tomorrow
-    var isTomorrow: Bool {
-        let calendar = Calendar.autoupdatingCurrent
-        return calendar.isDateInTomorrow(self)
-    }
-}
-
-
 extension UITableViewCell {
     static var reuseIdentifier: String { NSStringFromClass(self).components(separatedBy: ".").last! }
 }
 
 extension UICollectionReusableView {
     static var reuseIdentifier: String { NSStringFromClass(self).components(separatedBy: ".").last! }
-}
-
-extension String {
-    var localized: String {
-        return NSLocalizedString(self, comment: "")
-    }
 }
 
 extension UIImage {
